Add files via upload
In this I built and evaluated a text classifier which detects fake news. provided with a dataset of over 10,000 statements on current affairs which were annotated with labels for different degrees of fakeness, ranging from ‘true’ to ‘pants on fire’. Support vector machine was used which is a discriminative classifer.